+++ src/feedmark/tests.py
@@ -184,19 +184,19 @@ class TestFeedmarkCommandLine(unittest.TestCase):
         self.assertDictEqual(data, {
             "2 Llamas Spotted Near Mall": {
                 "anchor": "2-llamas-spotted-near-mall",
-                "filename": "eg/Recent Llama Sightings.md"
+                "filenames": ["eg/Recent Llama Sightings.md"],
             },
             "A Possible Llama Under the Bridge": {
                 "anchor": "a-possible-llama-under-the-bridge",
-                "filename": "eg/Recent Llama Sightings.md"
+                "filenames": ["eg/Recent Llama Sightings.md"],
             },
             "Llamas: It's Time to Spot Them": {
                 "anchor": "llamas-its-time-to-spot-them",
-                "filename": "eg/Recent Llama Sightings.md"
+                "filenames": ["eg/Recent Llama Sightings.md"],
             },
             "Maybe sighting the llama": {
                 "anchor": "maybe-sighting-the-llama",
-                "filename": "eg/Ancient Llama Sightings.md"
+                "filenames": ["eg/Ancient Llama Sightings.md"],
             }
         })
 
@@ -208,25 +208,6 @@ class TestFeedmarkCommandLine(unittest.TestCase):
         self.assertDictEqual(data, {
             "2 Llamas Spotted Near Mall": {
                 "anchor": "2-llamas-spotted-near-mall",
-                "filename": "eg/Recent Llama Sightings.md"
-            },
-            "A Possible Llama Under the Bridge": {
-                "anchor": "a-possible-llama-under-the-bridge",
-                "filename": "eg/Recent Llama Sightings.md"
-            },
-            "Llamas: It's Time to Spot Them": {
-                "anchor": "llamas-its-time-to-spot-them",
-                "filename": "eg/Referenced Llama Sightings.md"
-            },
-        })
-
-    def test_output_index_with_overlap(self):
-        # Both of these files contain an entry called "Llamas: It's Time to Spot Them"
-        main(['eg/Recent Llama Sightings.md', 'eg/Referenced Llama Sightings.md', '--output-multi-refdex'])
-        data = json.loads(sys.stdout.getvalue())
-        self.assertDictEqual(data, {
-            "2 Llamas Spotted Near Mall": {
-                "anchor": "2-llamas-spotted-near-mall",
                 "filenames": [
                     "eg/Recent Llama Sightings.md",
                 ]
@@ -235,7 +216,7 @@ class TestFeedmarkCommandLine(unittest.TestCase):
                 "anchor": "a-possible-llama-under-the-bridge",
                 "filenames": [
                     "eg/Recent Llama Sightings.md",
-                ]
+                ],
             },
             "Llamas: It's Time to Spot Them": {
                 "anchor": "llamas-its-time-to-spot-them",
@@ -243,6 +224,25 @@ class TestFeedmarkCommandLine(unittest.TestCase):
                     "eg/Recent Llama Sightings.md",
                     "eg/Referenced Llama Sightings.md"
                 ]
+            },
+        })
+
+    def test_output_refdex_with_overlap_forcing_single_filename(self):
+        # Both of these files contain an entry called "Llamas: It's Time to Spot Them"
+        main(['eg/Recent Llama Sightings.md', 'eg/Referenced Llama Sightings.md', '--output-refdex', '--output-refdex-single-filename'])
+        data = json.loads(sys.stdout.getvalue())
+        self.assertDictEqual(data, {
+            "2 Llamas Spotted Near Mall": {
+                "anchor": "2-llamas-spotted-near-mall",
+                "filename": "eg/Recent Llama Sightings.md",
+            },
+            "A Possible Llama Under the Bridge": {
+                "anchor": "a-possible-llama-under-the-bridge",
+                "filename": "eg/Recent Llama Sightings.md",
+            },
+            "Llamas: It's Time to Spot Them": {
+                "anchor": "llamas-its-time-to-spot-them",
+                "filename": "eg/Referenced Llama Sightings.md",
             },
         })